Make scopes optional
This also changes how we track scopes; we now defer to whatever the
deepest scope is at the time the lexeme is sliced, embracing the
possibility that it may be a None value. The previous last_scope
tracking didn't work across lines, exluding whole blocks of text that
didn't trigger a scope change on a new line iteration.